Added ch2p3, which was already solved in my implementation of linked list.
