+++ go/src/chapter2/linked_list.go
@@ -42,26 +42,6 @@ func (ll *DoublyLinkedList) Insert(val int) {
 	}
 }
 
-func (ll *DoublyLinkedList) Remove(index int) {
-	ll.removeNode(ll.getNode(index))
-}
-
-func (ll *DoublyLinkedList) removeNode(node *node) {
-	if ll.head == node {
-		ll.head = node.next
-	}
-	if ll.tail == node {
-		ll.tail = node.prev
-	}
-	if node.prev != nil {
-		node.prev.next = node.next
-	}
-	if node.next != nil {
-		node.next.prev = node.prev
-	}
-	ll.count--
-}
-
 func (ll *DoublyLinkedList) getNode(index int) *node {
 	node := ll.head
 	for i := 0; i < index; i++ {
